% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteger.R
\name{calcInteger}
\alias{calcInteger}
\title{Calculates the integer copy number profile for each single-cell}
\usage{
calcInteger(scCNA, method, ploidy_value = NULL)
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{method}{Method used to calculate integer profiles. Options: 'fixed'.}

\item{ploidy_value}{If method of choice is 'fixed' a ploidy value should be provided}
}
\value{
The scCNA object with an assay slot named 'integer' that contains
a data frame with cells as columns and integerized segments as rows.
}
\description{
Calculates the integer copy number profile for each single-cell
}
\details{
\item{fixed:} When method argument is set to 'fixed' copykit extracts the segment means
from the scCNA object and multiplies those means by the value provided in the
argument ploidy_value.
}
