% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatios.R
\name{calcRatios}
\alias{calcRatios}
\title{Calculates the ratios from a matrix of counts}
\usage{
calcRatios(scCNA, assay = "vst", fun = "median")
}
\arguments{
\item{scCNA}{The scCNA object}

\item{assay}{The assay that will be used}

\item{fun}{Character. Function used to calculate the ratios, defaults to "median"}
}
\value{
A ratio matrix within the slot \link{\code{assay(scCNA, 'ratios')}} can be accessed with \link{\code{copykit::ratios(scCNA)}}.
}
\description{
Calculates the ratios from a matrix of counts
}
