% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{igraph-class}
\alias{igraph-class}
\alias{CopyKit-class}
\alias{.CopyKit}
\alias{CopyKit}
\title{Placeholder for the igraph class}
\usage{
CopyKit(
  consensus = data.frame(),
  phylo = structure(list(), class = "phylo"),
  consensusPhylo = structure(list(), class = "phylo"),
  distMat = dist(matrix(0, 0, 0)),
  graph = igraph::graph.empty(),
  ...
)
}
\arguments{
\item{consensus}{A data frame with the consensus information.}

\item{phylo}{A phylo object with a phylogenetic tree.}

\item{consensusPhylo}{A phylo object with a phylogenetic consensus tree.}

\item{graph}{A graph object with a graph made from the umap data.}
}
\value{
A CopyKit class object.
}
\description{
S4 Class that extends the Bioconductor SingleCellExperiment class to hold
single cell copy number datasets.
}
\section{Slots}{

\describe{
\item{\code{phylo}}{Stores the single cell phylogenetic information with ape class
phylo.}

\item{\code{consensusPhylo}}{Stores the consensus phylogenetic information with
ape class phylo.}

\item{\code{distMat}}{Stores a distance matrix object used for graphs and heatmaps.}

\item{\code{graph}}{Stores an igraph object for network based clustering.}

\item{\code{consensus}}{stores a consensus data frame from
\code{\link{calcConsensus}.}}
}}

\references{
The Bioconductor SingleCellExperiment Class
DOI: 10.18129/B9.bioc.SingleCellExperiment
}
\keyword{internal}
