% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CopyKit}
\alias{CopyKit}
\alias{.CopyKit}
\title{The CopyKit class}
\value{
A CopyKit class object.
}
\description{
S4 Class that extends the Bioconductor SingleCellExperiment class to hold
single cell copy number datasets.
}
\section{Slots}{

\describe{
\item{\code{phylo}}{Stores the single cell phylogenetic information with ape class
phylo.}

\item{\code{consensusPhylo}}{Stores the consensus phylogenetic information with
ape class phylo.}

\item{\code{distMat}}{Stores a distance matrix object used for graphs and heatmaps.}

\item{\code{graph}}{Stores an igraph object for network based clustering.}

\item{\code{consensus}}{stores a consensus data frame from
\code{\link{calcConsensus}.}}
}}

\references{
The Bioconductor SingleCellExperiment Class
DOI: 10.18129/B9.bioc.SingleCellExperiment
}
