% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/internals.R
\docType{data}
\name{hg38_grangeslist}
\alias{hg38_grangeslist}
\alias{hg19_rg}
\alias{copykit_obj_rle}
\alias{copykit_obj_filt_rle}
\alias{copykit_obj_filt_umap}
\alias{hg19_genes}
\alias{hg38_genes}
\alias{copykit_example}
\alias{copykit_example_filtered}
\alias{mock_bincounts}
\title{hg38_grangeslist}
\format{
An object of class \code{CompressedGRangesList} of length 8.

An object of class \code{CompressedGRangesList} of length 8.

An object of class \code{list} of length 790.

An object of class \code{list} of length 1502.

An object of class \code{matrix} (inherits from \code{array}) with 790 rows and 2 columns.

An object of class \code{GRanges} of length 23056.

An object of class \code{GRanges} of length 30909.
}
\source{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)

library(TxDb.Hsapiens.UCSC.hg38.knownGene)
}
\usage{
hg38_grangeslist

hg38_grangeslist

copykit_obj_filt_rle

copykit_obj_rle

copykit_obj_filt_umap

hg19_genes

hg38_genes

copykit_example()

copykit_example_filtered()

mock_bincounts(
  ncells = 30,
  ncells_diploid = 5,
  position_gain = 4900:5493,
  position_del = 6523:7056,
  genome = "hg38",
  resolution = "220kb"
)
}
\arguments{
\item{ncells}{A numeric with the total number of cells to simulate.}

\item{ncells_diploid}{A numeric with the number of diploid cells to simulate}

\item{position_gain}{A vector with the index of the bin counts in which
chromosomal gains will be added.}

\item{position_del}{A vector with the index of the bin counts in which
chromosomal deletions will be placed.}

\item{genome}{The assembly genome information to add to the metadata.}

\item{resolution}{The resolution of the scaffold to add to the object
metadata}
}
\value{
Contains a GrangesList object with the scaffolds for each of the
resolutions used by runVarbin, runCountReads and runSegmentation on the
hg38 genome assembly.

Contains a GrangesList object with the scaffolds for each of the
resolutions used by runVarbin, runCountReads and runSegmentation on the hg19
genome assembly.

A CopyKit object with data from BL1 sample from the CopyKit
manuscript.

A CopyKit object with data from BL1 sample from the CopyKit
manuscript with diploid and noise cells removed.

A CopyKit object with simulated bincounts
}
\description{
hg38_grangeslist

hg19_rg

copykit_obj_rle

copykit_obj_filt_rle

copykit_obj_filt_umap

hg19_genes

hg38_genes

CopyKit Example

CopyKit Example

mock_bincounts
}
\details{
Contains a Rle object with the segment ratios values of the copykit
BL1 example dataset. It is used by the functions copykit_example

Contains a Rle object with the segment ratios values of the copykit
BL1 example filtered dataset. It is used by the functions
copykit_example_filtered()

Contains the umap reduced dimension for the BL1 dataset as generated
for the CopyKit manuscript

Contains the GrangesObject for the genomic positions of genes in the
hg19 genome assembly

Contains the GrangesObject for the genomic positions of genes in the
hg38 genome assembly

This function is largely used for examples of CopyKit functions.

This function returns a CopyKit object with the unfiltered dataset
of the sample BL1. This sample was processed with ACT and is a liver
metastasis from a primary breast cancer patient.

The CopyKit object contain only the segment ratio mean of each single cell
and, therefore, is not compatible with the functions from the runVarbin
module.

This function is largely used for examples of CopyKit functions.

This function returns a CopyKit object with the unfiltered dataset
of the sample BL1. This sample was processed with ACT and is a liver
metastasis from a primary breast cancer patient.

The CopyKit object contain only the segment ratio mean of each single cell
and, therefore, is not compatible with the functions from the runVarbin
module.
}
\keyword{internal}
