% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBreakpoints.R
\name{.countBreakpoints}
\alias{.countBreakpoints}
\title{countBreakpoints}
\usage{
.countBreakpoints(scCNA)
}
\arguments{
\item{scCNA}{}
}
\value{
The scCNA object with a column of breakpoint counts added to colData.
}
\description{
Considers changes in the segment ratios as breakpoints.
Counts the breakpoints for each chromosome arm separately.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- .countBreakpoints(copykit_obj)

}
\keyword{internal}
