% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAneuploidCells.R
\name{findAneuploidCells}
\alias{findAneuploidCells}
\title{findAneuploidCells}
\usage{
findAneuploidCells(
  scCNA,
  assay = "segment_ratios",
  resolution = "auto",
  remove_XY = TRUE,
  simul = TRUE,
  seed = 17
)
}
\arguments{
\item{scCNA}{The CopyKit object}

\item{assay}{String with the name of the assay to pull data from to find
normal cells.}

\item{resolution}{A numeric scalar used as threshold to detect normal cells.}

\item{remove_XY}{A boolean that removes chrX and chrY from the analysis.}

\item{simul}{A boolean that if TRUE adds a simulated normal dataset to boost
identifying normal cells in datasets with small proportions of normal cells.}

\item{seed}{Seed passed on to reproduce simulated CV of normal cells.}
}
\value{
information is added to \code{\link[SummarizedExperiment]{colData}}
in a columns named 'is_aneuploid' being TRUE if a cell is detected as
aneuploid and FALSE if the cell is detected as euploid.
}
\description{
Find cells that are not aneuploid in the dataset.
}
\details{
performs a sample-wise calculation of the segment means coefficient
 of variation and fits a Gaussian mixture model to the observed distribution
 from all cells. To increase the sensitivity of the model, the expected
 distribution of the coefficient of variation for euploid cells is simulated
 for a thousand cells (mean = 0, sd = 0.01). This way, CopyKit can adequately
 detect euploid cells even in datasets with limited amounts of euploid cells
 and guarantees that no aneuploid cell will be removed from datasets without
 any euploid cells. The distribution with the smallest CV
 is assumed originate from normal cells. Cells are classified as euploid
 if they have a coefficient of variance smaller than the mean plus five times
 the standard deviation of the normal cell distribution.
}
\examples{
copykit_obj <- copykit_example()
copykit_obj <- findAneuploidCells(copykit_obj)
}
