% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNormalCells.R
\name{findNormalCells}
\alias{findNormalCells}
\title{identifies possible normal cells in the dataset based on coefficient of variation}
\usage{
findNormalCells(scCNA)
}
\arguments{
\item{scCNA}{scCNA object}
}
\value{
Adds is_normal column to the scCNA metadata. Can be accessed with colData(scCNA)
}
\description{
identifies possible normal cells in the dataset based on coefficient of variation
}
