% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalK.R
\name{findOptimalK}
\alias{findOptimalK}
\title{Finds the Optimal K value to be used for subclone clustering}
\usage{
findOptimalK(
  scCNA,
  k_range = 10:30,
  method = "hdbscan",
  seed = 17,
  B = 100,
  n_threads = parallel::detectCores()/4
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{k_range}{Range of values to be tested. Defaults to 10:30}

\item{method}{Method which where the values will be tested. Only "hdbscan" available.}

\item{seed}{Seed (Defaults to 17).}

\item{B}{Number of bootstrapping. Defaults to 100. Higher values yield better results at a cost of performance}

\item{n_threads}{Number of threads. Passed to `parallel::mclapply`. As default it uses 1/4 of the detected cores available.}
}
\value{

}
\description{
Finds the Optimal K value to be used for subclone clustering
}
