% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSuggestedK.R
\name{findSuggestedK}
\alias{findSuggestedK}
\title{Finds the suggested K value to be used for subclone clustering}
\usage{
findSuggestedK(
  scCNA,
  embedding = "umap",
  k_range = 7:sqrt(ncol(segment_ratios(scCNA))),
  method = "hdbscan",
  seed = 17,
  B = 100,
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{embedding}{String with the name of the reducedDim to pull data from.}

\item{k_range}{Range of values to be tested.
Defaults to 7 to the sqrt of the number of cells}

\item{method}{Method which where the values will be tested.
Only "hdbscan" available.}

\item{seed}{Seed (Defaults to 17).}

\item{B}{Number of bootstrapping. Defaults to 100.
Higher values yield better results at a cost of performance}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
Adds a table with the mean jaccard coefficient of clusters for each
tested k and the suggested k value to be used for clustering to the scCNA metadata.
}
\description{
Finds the suggested K value to be used for subclone clustering
}
