% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableGenes.R
\name{findVariableGenes}
\alias{findVariableGenes}
\title{findVariableGenes}
\usage{
findVariableGenes(scCNA, genes, assay = "logr", top_n = 50)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{genes}{A vector of strings containing the HUGO Symbol for the gene
of interest.}

\item{assay}{String with the name of the assay to pull data with the copy
number states for each gene.}

\item{top_n}{A numeric defining how many variable genes will be returned.}
}
\value{
A string vector with the HUGO genes in decreasing order of importance
stored to the \code{\link[S4Vectors]{metadata}}.
}
\description{
Find the most variable genes in the dataset.
}
\details{
\code{findVariableGenes} Runs \code{\link[stats]{prcomp}} to the
copy number states of the genes from the provided gene list and returns
the one that have the largest absolute variance as assesed by the
loadings of the first principal component.

The resulting list of genes is stored within the metadata of the scCNA
object and can be accessed with \code{\link[S4Vectors]{metadata}}.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findVariableGenes(copykit_obj,
               genes = c("FHIT", "PTEN", "FOXO1", "BRCA1"))

}
