% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCopyPlot.R
\name{geneCopyPlot}
\alias{geneCopyPlot}
\title{Plot gene segment ratios}
\usage{
geneCopyPlot(scCNA, genes, geom = "swarm", label = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{genes}{Vector containing the HUGO Symbol for the genes of interest.}

\item{geom}{Which geom should be used for plotting, options are "violin" or "swarm". Defaults to "swarm"}

\item{label}{Color by an element of metadata. Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}}
}
\value{
A violin plot with the segment ratios for the genes of interest.
}
\description{
geneCopyPlot is a way to visualize the number of copies for different genes across all the cells. It outputs a violin plot with the desired genes.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
