% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferMrca.R
\name{inferMrca}
\alias{inferMrca}
\title{inferMrca}
\usage{
inferMrca(scCNA, value = 1)
}
\arguments{
\item{scCNA}{the scCNA object}

\item{value}{A numeric value used to compare the profiles to infer the mrca,
usually equal to the mean segment ratio of cells (value = 1) or the average
copy number of the cells}
}
\value{
Returns a numeric vector added to the \code{\link[S4Vectors]{metadata}}
of the scCNA object named `inferred_mrca`
}
\description{
From the consensus matrix it infers a Most Recent Common Ancestral (MRCA)
across all groups.
}
\details{
Calculates the MRCA by inferring, for every bin, the value across
all groups that is closest to the number supplied in the argument value.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
copykit_obj <- calcConsensus(copykit_obj)
copykit_obj <- inferMrca(copykit_obj)

}
