% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNorm.R
\name{logNorm}
\alias{logNorm}
\title{logNorm()}
\usage{
logNorm(
  scCNA,
  transform = c("log", "log2", "log10", "log1p"),
  assay = "segment_ratios",
  name = "logr"
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{transform}{String specifying the transformation to apply to the selected
assay.}

\item{assay}{String with the name of the assay to pull data from to run the
segmentation.}

\item{name}{String with the name for the target slot for the resulting
transformed counts.}
}
\value{
A data frame with log transformed counts inside the
\code{\link[SummarizedExperiment]{assay}} slot.
}
\description{
Computes a log transformation of the selected assay
}
\examples{
copykit_obj <- copykit_example()
copykit_obj <- logNorm(copykit_obj)

}
