% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsensusLine.R
\name{plotConsensusLine}
\alias{plotConsensusLine}
\title{plotConsensusLine}
\usage{
plotConsensusLine(scCNA)
}
\arguments{
\item{scCNA}{The scCNA object}
}
\value{
An interactive plot where different groups
can be selected for easier visualization
}
\description{
Opens an interactive app showing ratio line plots of the different consensus
groups
}
\examples{
\dontrun{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
copykit_obj <- calcConsensus(copykit_obj)
plotConsensusLine(copykit_obj)
}

}
