% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreq.R
\name{plotFreq}
\alias{plotFreq}
\title{plotFreq}
\usage{
plotFreq(
  scCNA,
  high_threshold = 1.3,
  low_threshold = 0.7,
  assay = "segment_ratios",
  label = NULL,
  geom = c("area", "line"),
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{high_threshold}{A numeric with the threshold above which events are
considered amplifications.}

\item{low_threshold}{A numeric with the threshold below which events are
considered deletions.}

\item{assay}{String with the name of the assay to pull data from to plot
the frequency plot.}

\item{label}{A string with the name of the columns from
\code{\link[SummarizedExperiment]{colData}} to separate each frequency plot.}

\item{geom}{A character with the desired geom}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
A ggplot object with a frequency plot
}
\description{
plotFreq
}
\details{
\code{plotFreq} retrieves the data from the desired assay and creates
an event matrix based on the high and low thresholds arguments. Values above
the high threshold will be classified as gains whereas values below are classified
as deletions. The resulting plot is a frequency plot where values above 0
represent the frequency of gains and values below 0 represent the frequency of
deletions.

If the argument 'label' is provided the frequency plot will be calculated
separately for each label. Labels can be any string column from
\code{\link[SummarizedExperiment]{colData}}

The following geoms are available:

\itemize{
\item{area}: If geom = 'area' an area plot with the frequency is plotted.
If the label argument is provided a different facet will be plotted for each
label.

\item{line}: If geom = 'line' a line plot with the frequency is plotted.
If the label argument lines are overlapped with different colors.

}
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
plotFreq(copykit_obj, label = 'subclones')

}
