% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Plot metrics}
\usage{
plotMetrics(scCNA, metric, label = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{metric}{A character indicating which elements of \code{colData()}
should be plotted.}

\item{label}{A character indicating which element of the \code{colData()} to
color the plots.}
}
\value{
A ggplot object with the metrics available in the metadata.
}
\description{
Provides a plot of the metrics to quickly assess
the quality and overall look of the copy number data.
}
\examples{


}
\author{
Darlan Conterno Minussi
}
