% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPca.R
\name{plotPca}
\alias{plotPca}
\title{plotPca}
\usage{
plotPca(scCNA, embedding = "PCA", label = NULL)
}
\arguments{
\item{scCNA}{The CopyKit object.}

\item{embedding}{String with the name of the reducedDim to pull data from.}

\item{label}{A string with the elements from
\code{\link[SummarizedExperiment]{colData}} to color the umap points.}
}
\value{
A ggplot object containing the reduced dimensions UMAP plot.
}
\description{
Plots PCA embedding stored in \code{\link[SingleCellExperiment]{reducedDim}}
slot.
}
\details{
A reduced dimension representation with UMAP in the slot
\code{\link[SingleCellExperiment]{reducedDim}} from the scCNA object.

Columns from \code{\link[SummarizedExperiment]{colData}} can
be used as an argument for 'label' to color the points on the plot.
}
\examples{
set.seed(1000)
copykit_obj <- copykit_example_filtered()[,sample(300)]
copykit_obj <- runPca(copykit_obj)

plotPca(copykit_obj)

copykit_obj <- findClusters(copykit_obj)

plotPca(copykit_obj, label = "subclones")

}
\author{
Darlan Conterno Minussi
}
