% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatio.R
\name{plotRatio}
\alias{plotRatio}
\title{Plot ratio}
\usage{
plotRatio(scCNA, sample_name = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{sample_name}{Optional character vector with the name of the sample to
be visualized}
}
\value{
Opens an app for interactive visualization of the ratio plots where
the desired cell can be selected. If a 'sample_name' is provided, returns a ggplot
object with the ratio plot for the selected cell.
}
\description{
plotRatio allows for a visualization of the segment ratios together with the ratios.
It is helpful to observe the fit of the segmentation to the data.
}
\details{
plotRatio will return a ratio plot for the selected cell. If
\code{\link{calcInteger}} was run and an 'integer' assay exists, plotRatio will
color the ratio dots based on the copy number integer state of the segment
automatically.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
