% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatioPlot.R
\name{plotRatioPlot}
\alias{plotRatioPlot}
\title{Plot ratio plot}
\usage{
plotRatioPlot(scCNA, sample_name = NULL, interactively = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{sample_name}{character vector with the name of the sample to be visualized}
}
\value{
A violin plot with the segment ratios for the genes of interest.
}
\description{
plotRatioPlot allows for a visualization of the segment ratios together with the ratios.
It is helpful to observe the fit of the segmentation to the data.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
