% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSuggestedK.R
\name{plotSuggestedK}
\alias{plotSuggestedK}
\title{plotSuggestedK}
\usage{
plotSuggestedK(scCNA, geom = c("boxplot", "tile", "dotplot"))
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{geom}{A character with the geom to be used for plotting.}
}
\value{
A ggplot2 object with the plot of different tested k values and their
jaccard similarity for each subclone
}
\description{
Uses the information from \code{\link{findSugggestedK}} to plot the values
of jaccard similarity from the tested k range on \code{\link{findSugggestedK}}.
}
\details{
\code{\link{plotSuggestedK}} access the \code{\link[S4Vectors]{metadata}}
element suggestedK_df that is saved to the scDNA object after running
\code{\link{findSugggestedK}}. The dataframe is used for plotting either a
heatmap, when the argument geom = 'tile', or a dotplot when argument geom =
'dotplot' or a boxplot when geom = 'boxplot'.

\itemize{
\item{geom = 'boxplot':} With geom boxplot the red dots represent the mean
jaccard similarity for each assesed k.
}
}
