% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariableGenes.R
\name{plotVariableGenes}
\alias{plotVariableGenes}
\title{plotVariableGenes}
\usage{
plotVariableGenes(scCNA, genes, assay = "segment_ratios", top_n = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{genes}{A vector of strings containing the HUGO Symbol for the gene
of interest.}

\item{assay}{String with the name of the assay to pull data from to calculate
integers.}

\item{top_n}{A numeric that if provided returns the top N variable genes.}
}
\value{

}
\description{
Uses PCA to detect the most variable genes.
}
