% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVarbinCNA.R
\name{readVarbinCNA}
\alias{readVarbinCNA}
\title{Load data from Copy Number Experiment}
\usage{
readVarbinCNA(
  dir,
  remove_Y = FALSE,
  genome_version = c("hg19"),
  bin_size = c("200k", "100k"),
  clean_names = TRUE
)
}
\arguments{
\item{dir}{A path for the output of the copy number pipeline.}

\item{remove_Y}{(default == FALSE) If set to TRUE, removes information from the chrY from the dataset.}

\item{genome_version}{Name of the genome assembly. Default: 'hg19'.}

\item{bin_size}{The resolution of the VarBin method. Default: '200k'. Available options: '100k', '200k'.}

\item{clean_names}{Clean sample names using \code{\link{janitor::clean_names}}.}
}
\value{
Segment ratios can be acessed with \code{copykit::segment_ratios}.

Ratios can be acessed with \code{copykit::ratios}.

Bin counts can be acessed with \code{copykit::bin_counts}.

Genomic ranges can be acessed with \code{SummarizedExperiment::rowRanges()}
}
\description{
Creates an S4 class scCNA from the output directory of the copy number pipeline.
readVarbinCNA searches for the uber*.seg uber.bin and uber.ratio files resulting from the Varbin copy number pipeline in the provided directory.
The scCNA object contains the segment ratios, ratios and bincounts within the assay slot. where each bin is row and each sample (cell) is a column.
Genomic ranges are stored in a GRanges object containing chromosome number, start coordinate, end cordinate and absolute genomic position. Each row represents the coordinates for one bin.
}
\examples{

}
\author{
Darlan Conterno Minussi, Yun Yan
}
