% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runConsensusPhylo.R
\name{runConsensusPhylo}
\alias{runConsensusPhylo}
\title{runConsensusPhylo}
\usage{
runConsensusPhylo(scCNA)
}
\arguments{
\item{consensus_df}{The consensus data frame from the segmented dataset}

\item{clusters}{The results from clustering}

\item{ploidy_VAL}{The inferred ploidy value. Will be used to scale the data}

\item{rotate_nodes}{Nodes to be rotated}

\item{plot}{Boolean. Prints the plot if true}
}
\value{

}
\description{
Runs a minimal evolution tree algorithm for the consensus data frame
}
