% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDistMat.R
\name{runDistMat}
\alias{runDistMat}
\title{Run distance matrix calculations}
\usage{
runDistMat(scCNA, metric = "euclidean", n_threads = 1)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{metric}{distance metric passed to calculate the distance matrix.}

\item{n_threads}{Number of threads used to calculate the distance matrix.
Passed to `amap::Dist`.}
}
\value{
A distance matrix in the slot \code{distMat} from scCNA object.
Access the distance matrix with: \code{distMat(scCNA, withDimnames = TRUE)}
}
\description{
Performs distance matrix calculations that can be downstream used for
hierarchical clustering or phylogenetic analysis. Uses \code{amap::Dist()}
in order to parallelize distance calculations.
}
\examples{
copykit_obj <- copykit_example_filtered()[,1:10]
copykit_obj <- runDistMat(copykit_obj)
}
\author{
Darlan Conterno Minussi
}
