% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPca.R
\name{runPca}
\alias{runPca}
\title{runPca()}
\usage{
runPca(scCNA, assay = "logr", name = "PCA", scale = FALSE, ...)
}
\arguments{
\item{scCNA}{The CopyKit object.}

\item{assay}{String with the name of the assay to pull data from to make the
embedding.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{scale}{A logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place.}

\item{...}{Additional parameters passed to \code{\link[stats]{prcomp}}.}
}
\value{
A reduced dimension representation with pca in the slot
\code{reducedDim} from scCNA object. Access reduced dimensions slot with:
\code{reducedDim(scCNA, 'PCA', withDimnames = FALSE)}
}
\description{
Creates a pca embedding using the package uwot from the segment ratios values
}
\author{
Darlan Conterno Minussi
}
