% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmap.R
\name{runUmap}
\alias{runUmap}
\title{Creates UMAP embedding}
\usage{
runUmap(scCNA, assay = "logr", seed = 17, name = "umap", ...)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{assay}{String with the name of the assay to pull data from to make the embedding.}

\item{seed}{Sets a seed for the pseudorandom number generator.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{...}{Additional parameters passed to \code{uwot::umap}.}
}
\value{
A reduced dimension representation with UMAP in the slot \code{reducedDim} from scCNA object. Access reduced dimensions slot with: \code{SingleCellExperiment::reducedDim(scCNA, 'umap', withDimnames = FALSE)}
}
\description{
Creates a umap embedding using the package uwot from the segment ratios values
}
\author{
Darlan Conterno Minussi
}
