% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVarbin.R
\name{runVarbin}
\alias{runVarbin}
\title{Read BAM file and perform varbin analysis}
\usage{
runVarbin(
  dir,
  genome = "hg38",
  bin_size = "200kb",
  remove_Y = FALSE,
  min_reads = 1e+05,
  BPPARAM = bpparam()
)
}
\arguments{
\item{dir}{A path for the directory containing .BAM files from short-read sequencing.}

\item{genome}{Name of the genome assembly. Default: 'hg38'.}

\item{bin_size}{The resolution of the VarBin method. Default: '200kb'.}

\item{remove_Y}{(default == FALSE) If set to TRUE, removes information from the chrY from the dataset.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
Segment ratios can be accessed with \code{copykit::segment_ratios}.

Ratios can be accessed with \code{copykit::ratios}.

Bin counts can be accessed with \code{copykit::bin_counts}.

Genomic ranges can be accessed with \code{SummarizedExperiment::rowRanges()}
}
\description{
runVarbin performs the variable binning (VarBin) algorithm to a series of BAM files resulting from short-read sequencing.
}
\examples{


}
\author{
Darlan Conterno Minussi
}
