% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVst.R
\name{runVst}
\alias{runVst}
\title{Variance Stabilizing Transformation}
\usage{
runVst(scCNA)
}
\arguments{
\item{scCNA}{The scCNA object}
}
\value{
A slot into the scCNA object containing the variance stabilized matrix. Can be accessed with \code{assay(scCNA, 'vst')}
}
\description{
Performs variance stabilization by using the Freeman-Tukey transformation
}
