% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{segment_ratios,CopyKit-method}
\alias{segment_ratios,CopyKit-method}
\title{Internal CopyKit functions}
\usage{
\S4method{segment_ratios}{CopyKit}(x, withDimnames = TRUE)
}
\description{
Methods to get or set internal fields from the CopyKit class
}
