% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{methods}
\name{segment_ratios,scCNA-method}
\alias{segment_ratios,scCNA-method}
\title{Internal CopyKit functions}
\usage{
\S4method{segment_ratios}{scCNA}(x, withDimnames = TRUE)
}
\description{
Methods to get or set internal fields from the scCNA class
}
