% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConsensus.R
\name{calcConsensus}
\alias{calcConsensus}
\title{Calculate a consensus matrix of segment ratios based on metadata}
\usage{
calcConsensus(
  scCNA,
  consensus_by = "subclones",
  n_threads = parallel::detectCores()/4
)
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{consensus_by}{The column from metadata that will be used
to isolate the cells by factor and calculate the consensus.}

\item{n_threads}{Number of threads used to calculate the distance matrix.
Passed to `parallel::mclapply`.
As default it uses 1/4 of the detected cores available.}
}
\value{

}
\description{
Calculate a consensus matrix of segment ratios based on metadata
}
