% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConsensus.R
\name{calcConsensus}
\alias{calcConsensus}
\title{Calculate a consensus matrix of segment ratios based on metadata}
\usage{
calcConsensus(
  scCNA,
  assay = "segment_ratios",
  consensus_by = "subclones",
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{assay}{assay from the scCNA object wherein the consensus matrix is
calculated}

\item{consensus_by}{The column from metadata that will be used
to isolate the cells by factor and calculate the consensus.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{

}
\description{
Calculate a consensus matrix of segment ratios based on metadata
}
