% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteger.R
\name{calcInteger}
\alias{calcInteger}
\title{Calculates the integer copy number profile for each single-cell}
\usage{
calcInteger(
  scCNA,
  assay = "segment_ratios",
  method = "fixed",
  ploidy_value = NULL,
  name = "integer",
  penalty = 25,
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{assay}{String with the name of the assay to pull data from to calculate integers.}

\item{method}{Method used to scale the ratio values to integer.}

\item{ploidy_value}{If method of choice is 'fixed' a ploidy value should be provided.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{penalty}{An integer passed on to scquantum::ploidy.inference() penalty argument}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
The scCNA object with an assay slot named 'integer' that contains
a data frame with cells as columns and integerized segments as rows.
}
\description{
Calculates the integer copy number profile for each single-cell
}
\details{
\itemize{
\item{fixed:} When method argument is set to 'fixed' copykit extracts the segment means
from the scCNA object and multiplies those means by the value provided in the
argument ploidy_value.
}
}
