% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteger.R
\name{calcInteger}
\alias{calcInteger}
\title{calcInteger()}
\usage{
calcInteger(
  scCNA,
  assay = c("bincounts", "smoothed_bincounts", "segment_ratios"),
  method = "fixed",
  ploidy_value = NULL,
  name = "integer",
  penalty = 25,
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{The CopyKit object.}

\item{assay}{String with the name of the assay to pull data from to calculate
integers.}

\item{method}{Method used to scale the ratio values to integer.}

\item{ploidy_value}{If method of choice is 'fixed' a ploidy value should be
provided.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{penalty}{An integer passed on to scquantum::ploidy.inference()
penalty argument}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
The CopyKit object with an assay slot named 'integer' that contains
a data frame with cells as columns and integerized segments as rows. And, in
case of method = 'scquantum' CopyKit adds  two new elements to \code{colData}
 named 'ploidy' and 'ploidy_confidence' with the respective values for
 each cell.
}
\description{
Calculates the integer copy number profile for each single cell
}
\details{
CopyKit support the following methods for calculating integer copy number
matrices
\itemize{
\item{fixed:} When method argument is set to 'fixed' copykit extracts the
segment means from the scCNA object and multiplies those means by the value
provided in the argument ploidy_value.
}

\item{scquantum:} When the method argument is set to 'scquantum', CopyKit
applies \code{\link[scquantum]{ploidy.inference}} function to perform a
sample wise calculation returning the estimated compuational ploidy for
every single cell
}
\examples{
copykit_obj <- mock_bincounts(ncells_diploid = 0, ncells = 10)
copykit_obj <- calcInteger(copykit_obj, method = "scquantum")
}
