% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInteger.R
\name{calcInteger}
\alias{calcInteger}
\title{Calculates the integer copy number profile for each single-cell}
\usage{
calcInteger(scCNA, method = "fixed", ploidy_value = NULL, slot = "integer")
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{method}{Method used to scale the ratio values to integer.}

\item{ploidy_value}{If method of choice is 'fixed' a ploidy value should be provided.}

\item{slot}{Name of the resulting slot where integer matrix will be saved.}
}
\value{
The scCNA object with an assay slot named 'integer' that contains
a data frame with cells as columns and integerized segments as rows.
}
\description{
Calculates the integer copy number profile for each single-cell
}
\details{
\itemize{
\item{fixed:} When method argument is set to 'fixed' copykit extracts the segment means
from the scCNA object and multiplies those means by the value provided in the
argument ploidy_value.
}
}
