% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatios.R
\name{calcRatios}
\alias{calcRatios}
\title{Calculates the ratios from a matrix of counts}
\usage{
calcRatios(scCNA, assay = "ft", fun = "median")
}
\arguments{
\item{scCNA}{The scCNA object}

\item{assay}{The assay that will be used}

\item{fun}{Character. Function used to calculate the ratios.
Defaults to "median"}
}
\value{
A ratio matrix within the slot \code{assay(scCNA, 'ratios')}
can be accessed with \code{copykit::ratios(scCNA)}.
}
\description{
Calculates the ratios from a matrix of counts
}
