% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatios.R
\name{calcRatios}
\alias{calcRatios}
\title{Calculates the ratios from a matrix of counts}
\usage{
calcRatios(
  scCNA,
  assay = c("ft", "bin_counts", "smoothed_bincounts"),
  fun = c("mean", "median")
)
}
\arguments{
\item{scCNA}{The scCNA object}

\item{assay}{String with the name of the assay to pull data from to calculate
the ratios.}

\item{fun}{A string indicating the summarizing function to be used.}
}
\value{
A ratio matrix within the slot assay(scCNA, 'ratios')
can be accessed with \code{ratios}.
}
\description{
Calculates the ratios from a matrix of counts
}
\details{
Calculates a sample-wise normalization of the selected assay by the
mean bin counts returns ratios where a value of 1 corresponds to the neutral
copy number state of the sample
}
