% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R, R/internals.R
\docType{class}
\name{CopyKit-class}
\alias{CopyKit-class}
\alias{.CopyKit}
\alias{segment_ratios}
\alias{ratios}
\alias{bincounts}
\alias{segment_ratios,CopyKit-method}
\title{@export
 @importMethodsFrom SummarizedExperiment colData
 @importMethodsFrom SingleCellExperiment SingleCellExperiment
 @importClassesFrom SummarizedExperiment RangedSummarizedExperiment SingleCellExperiment
 @importClassesFrom S4Vectors DataFrame SimpleList
A S4 class to store copy number assays with CopyKit.
Inherits from SingleCellExperiment}
\usage{
segment_ratios(x, ...)

ratios(x, ...)

bincounts(x, ...)

\S4method{segment_ratios}{CopyKit}(x, withDimnames = TRUE)
}
\value{
A CopyKit class object.
}
\description{
@export
 @importMethodsFrom SummarizedExperiment colData
 @importMethodsFrom SingleCellExperiment SingleCellExperiment
 @importClassesFrom SummarizedExperiment RangedSummarizedExperiment SingleCellExperiment
 @importClassesFrom S4Vectors DataFrame SimpleList
A S4 class to store copy number assays with CopyKit.
Inherits from SingleCellExperiment
}
\section{Slots}{

\describe{
\item{\code{phylo}}{Stores the single cell phylogenetic information with ape class
phylo.}

\item{\code{consensusPhylo}}{Stores the consensus phylogenetic information with
ape class phylo.}

\item{\code{distMat}}{Stores a distance matrix object used for graphs and heatmaps.}

\item{\code{graph}}{Stores an igraph object for network based clustering.}

\item{\code{consensus}}{stores a consensus data frame from \link{calcConsensus}.}
}}

