% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBreakpoints.R
\name{.countBreakpoints}
\alias{.countBreakpoints}
\title{Counting breakpoints from}
\usage{
.countBreakpoints(scCNA)
}
\arguments{
\item{scCNA}{}
}
\value{
The scCNA object with a column of breakpoint counts added to the colData.
}
\description{
Considers changes in the segment ratios as breakpoints.
Counts the breakpoints for each chromosome separately.
}
\keyword{internal}
