% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{filterCells}
\alias{filterCells}
\title{Filter noise cells}
\usage{
filterCells(scCNA, k = 5, resolution = 0.8, n_threads = 1)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{k}{K-nearest-neighbor, defaults to 5.}

\item{resolution}{Set's how strict the correlation cut off will be. Defaults to 0.8.}

\item{n_threads}{Number of parallel threads to calculate distances with \code{amap::Dist()}. Defaults to 1}
}
\value{
Adds a filtered cells label to the scCNA metadata. Cells that pass the filtering criteria receive the label "kept", whereas cells that do not pass the filtering criteria receive the label "removed".
}
\description{
filterCells uses a k-nearest-neighbor approach to remove cells with random CNA profiles, largely due to noise data. It calculates a correlation matrix and sets a resolution below which non neighbors will be classified as noise cells.
Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}
}
\examples{


}
\author{
Hua-Jun Wu
}
