% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{Find Clusters}
\usage{
findClusters(
  scCNA,
  method = "hdbscan",
  k_major = NULL,
  k_minor = NULL,
  seed = 17
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{method}{Which method should be used for clustering, options are "hdbscan" or "leiden". Defaults to "hdbscan".}

\item{k_major}{k-nearest-neighbor value that will be used to find the major clusters. Must be higher than k_minor (Defaults to 35).}

\item{k_minor}{k-nearest-neighbor value that will be used to find the minor clusters (Defaults to 21).}

\item{seed}{Seed passed on to leiden algorithm (Defaults to 17).}
}
\value{
Metadata cluster information that can be found in \code{SummarizedExperiment::colData(scCNA)$superclones} for the major clusters and \code{SummarizedExperiment::colData(scCNA)$subclones} for the minor clusters. Major clusters are named with capital letters whereas minor clusters are named with a numbers but as a character vector.
}
\description{
Search for clusters in the scCNA data by using a graph based approach. \code{findClusters()} builds an SNN graph of the k-nearest neighbors and attempts to find two different configuration of clusters. Major and minor subpopulations.
Major clusters are found by looking at the graph connected components, whereas the minor clusters use the Leiden algorithm to detect connected communities within the major clusters.
\code{findClusters()} generates the graph by using the UMAP embedding that can be obtained after running \code{runUmap()}.
}
\author{
Darlan Conterno Minussi
}
