% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{Find Clusters}
\usage{
findClusters(
  scCNA,
  method = "hdbscan",
  k_superclones = NULL,
  k_subclones = NULL,
  seed = 17
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{method}{Which method should be used for clustering,
options are "hdbscan" or "leiden". Defaults to "hdbscan".}

\item{k_superclones}{k-nearest-neighbor value.
Used to find the major clusters.}

\item{k_subclones}{k-nearest-neighbor value.
Used to find the minor clusters}

\item{seed}{Seed passed on to pseudorandom dependent functions (Defaults to 17).}
}
\value{
Metadata cluster information that can be found in
\code{SummarizedExperiment::colData(scCNA)$superclones}
for the major clusters and \code{SummarizedExperiment::colData(scCNA)$subclones}
for the minor clusters.
}
\description{
Search for clusters in the scCNA data by
using a graph based approach. \code{findClusters()}
builds an SNN graph of the k-nearest neighbors and
attempts to find two different configuration of clusters.
Major and minor subpopulations.
Major clusters are found by looking at the graph connected components,
 whereas the minor clusters use the hdbscan or leiden
  algorithm to detect connected communities within the major clusters.
\code{findClusters()} generates the graph by using the
UMAP embedding that can be obtained after running \code{runUmap()}.
}
\author{
Darlan Conterno Minussi
}
