% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSuggestedK.R
\name{findSuggestedK}
\alias{findSuggestedK}
\alias{hdbscanCBI}
\alias{leidenCBI}
\alias{louvainCBI}
\title{findSuggestedK}
\usage{
findSuggestedK(
  scCNA,
  embedding = "umap",
  k_range = NULL,
  method = c("hdbscan", "leiden", "louvain"),
  metric = c("median", "mean"),
  seed = 17,
  B = 200,
  BPPARAM = bpparam()
)

hdbscanCBI(data, minPts, diss = inherits(data, "dist"), ...)

leidenCBI(data, k, seed_leid, diss = inherits(data, "dist"), ...)

louvainCBI(data, k, seed_leid, diss = inherits(data, "dist"), ...)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{embedding}{String with the name of the reducedDim embedding to pull data from.}

\item{k_range}{A numeric range of values to be tested.}

\item{method}{A string with the method of clustering to be tested.}

\item{metric}{A string with the function to summarize the jaccard similarity
value from all clusters.}

\item{seed}{A numerical scalar with a seed value to be passed on to
\code{\link[uwot]{umap}}.}

\item{B}{A numeric with the number of bootstrapping iterations passed on to
\code{\link[fpc]{clusterboot}}. Higher values yield better results at a cost
of performance}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
Adds a table with the mean jaccard coefficient of clusters for each
tested k and the suggested k value to be used for clustering to
\code{\link[S4Vectors]{metadata}}
}
\description{
Performs a grid search over a range of k values to assess cluster stability.
}
\details{
Performs a grid-search over a range of k values and returns the value
that maximizes the jaccard similarity. Importantly, while this approach does
not guarantee optimal clustering, it provides a guide that maximizes cluster
stability.

The default tested range is from 7 to the square root of the number of cells
in the scCNA object. If sqrt(n_cells) is smaller than 7 a range of 5 to 15
is tested.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findSuggestedK(copykit_obj)

}
\references{
Hennig, C. (2007) Cluster-wise assessment of cluster stability.
Computational Statistics and Data Analysis, 52, 258-271.

Hennig, C. (2008) Dissolution point and isolation robustness: robustness
criteria for general cluster analysis methods.
Journal of Multivariate Analysis 99, 1154-1176.
}
\seealso{
\code{\link[fpc]{clusterboot}}

\code{\link{plotSuggestedK}}
}
\keyword{internal}
