% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{methods}
\name{CopyKit-internals}
\alias{CopyKit-internals}
\alias{segment_ratios}
\alias{segment_ratios,CopyKit-method}
\alias{ratios}
\alias{ratios,CopyKit-method}
\alias{bincounts}
\alias{bincounts,CopyKit-method}
\alias{consensus}
\alias{consensus,CopyKit-method}
\alias{consensus<-}
\alias{consensus<-,CopyKit-method}
\alias{phylo}
\alias{phylo,CopyKit-method}
\alias{phylo<-}
\alias{phylo<-,CopyKit-method}
\alias{consensusPhylo}
\alias{consensusPhylo,CopyKit-method}
\alias{consensusPhylo<-}
\alias{consensusPhylo<-,CopyKit-method}
\alias{distMat}
\alias{distMat,CopyKit-method}
\alias{distMat<-}
\alias{distMat<-,CopyKit-method}
\alias{graph}
\alias{graph,CopyKit-method}
\alias{graph<-}
\alias{graph<-,CopyKit-method}
\alias{show}
\alias{show,CopyKit-method}
\alias{\%!in\%}
\alias{superclones_pal}
\alias{subclones_pal}
\alias{find_scaffold_genes}
\alias{l2e.normal.sd}
\alias{overdispersion}
\alias{parCor}
\title{CopyKit internal functions.}
\usage{
\S4method{segment_ratios}{CopyKit}(x, withDimnames = TRUE)

\S4method{ratios}{CopyKit}(x, withDimnames = TRUE)

\S4method{bincounts}{CopyKit}(x, withDimnames = TRUE)

\S4method{consensus}{CopyKit}(x, withDimnames = TRUE)

\S4method{consensus}{CopyKit}(x) <- value

\S4method{phylo}{CopyKit}(x)

\S4method{phylo}{CopyKit}(x) <- value

\S4method{consensusPhylo}{CopyKit}(x)

\S4method{consensusPhylo}{CopyKit}(x) <- value

\S4method{distMat}{CopyKit}(x)

\S4method{distMat}{CopyKit}(x) <- value

\S4method{graph}{CopyKit}(x)

\S4method{graph}{CopyKit}(x) <- value

\S4method{show}{CopyKit}(object)

x \%!in\% table

superclones_pal()

subclones_pal()

find_scaffold_genes(scCNA, genes)

l2e.normal.sd(xs)

overdispersion(v)

parCor(x, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{x}{CopyKit object.}
}
\value{
a named vector of default colors for CopyKit superclones.

a named vector of default colors for CopyKit subclones.

A data frame with the gene HUGO gene symbol and the position on the
relevant scaffold from the varbin pipeline.

A numeric vector with least squares sd estimation.

A numerci vector with the estimation of the index of dispersion,
which is used when estimating standard errors for each segment mean

A matrix with the pairwise correlation from the segment ratio means.
}
\description{
This document establish setters and getters to facilitate access
to fields for the CopyKit class object. The functions provided here are
in addition to setters and getters available from the SingleCellExperiment
class
}
\section{Getters}{

\describe{
\item{\code{segment_ratios}:}{Returns a data frame of normalized segment
ratio means.}
\item{\code{ratios}:}{Returns a data frame of normalized ratio means.}
\item{\code{bincounts}:}{Returns a data frame of binned bincounts.}
\item{\code{consensus}:}{Returns a data frame of normalized segment
ratio means for the consensus matrix.}
\item{\code{phylo}:}{Returns a phylo class object with a phylogenetic tree.}
\item{\code{consensusPhylo}:}{Returns a phylo class object with a
phylogenetic tree from the consensus matrix.}
}
}

\author{
Alexander Davis

Alexander Davis

Junke Wang
}
\keyword{internal}
