% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlluvial.R
\name{plotAlluvial}
\alias{plotAlluvial}
\title{plotAlluvial}
\usage{
plotAlluvial(scCNA, label, label_colors = NULL)
}
\arguments{
\item{label}{A string with two or more  elements from \code{\link[SummarizedExperiment]{colData}}.}

\item{label_colors}{An optional named vector with the colors of each element from label.}
}
\value{
A ggplot object containing an alluvial plot from ggalluvial
}
\description{
Produces an alluvial plot from character elements of the metadata
}
