% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlluvial.R
\name{plotAlluvial}
\alias{plotAlluvial}
\title{plotAlluvial()}
\usage{
plotAlluvial(scCNA, label, label_colors = NULL, min_cells = NULL)
}
\arguments{
\item{scCNA}{The CopyKit object.}

\item{label}{A string with two or more  elements from \code{\link[SummarizedExperiment]{colData}}.}

\item{label_colors}{An optional named vector with the colors of each element
from label.}

\item{min_cells}{An optional numeric to filter stratum that do not reach
the minimum amount of cells.}
}
\value{
A ggplot object containing an alluvial plot from ggalluvial
}
\description{
Produces an alluvial plot from character elements of the metadata
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
colData(copykit_obj)$section <- stringr::str_extract(
    colData(copykit_obj)$sample,
    "(L[0-9]+L[0-9]+|L[0-9]+)"
)
plotAlluvial(copykit_obj, label = c("subclones", "section"))
}
