% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsensusLine.R
\name{plotConsensusLine}
\alias{plotConsensusLine}
\title{plotConsensusLine}
\usage{
plotConsensusLine(scCNA)
}
\arguments{
\item{scCNA}{The scCNA object}
}
\value{
An interactive plot where different groups
can be selected for easier visualization
}
\description{
Opens an interactive app showing ratio line plots of the different consensus
groups
}
\examples{
copykit_obj <- copykit_example()
copykit_obj <- findNormalCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$is_normal == "FALSE"]
copykit_obj <- filterCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$filtered == "kept"]
copykit_obj <- runUmap(copykit_obj)
copykit_obj <- findSuggestedK(copykit_obj)
copykit_obj <- findClusters(copykit_obj)
copykit_obj <- calcConsensus(copykit_obj)
plotConsensusLine(copykit_obj)

}
