% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsensusLine.R
\name{plotConsensusLine}
\alias{plotConsensusLine}
\title{plotConsensusLine}
\usage{
plotConsensusLine(scCNA)
}
\arguments{
\item{scCNA}{The scCNA object}
}
\value{
An interactive plot where different groups
can be selected for easier visualization
}
\description{
plotConsensusLine
}
