% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCopy.R
\name{plotGeneCopy}
\alias{plotGeneCopy}
\title{plotGeneCopy}
\usage{
plotGeneCopy(
  scCNA,
  genes,
  geom = c("swarm", "barplot", "violin"),
  label = NULL,
  facet = NULL,
  dodge.width = 0,
  assay = "segment_ratios"
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{genes}{A vector of strings containing the HUGO Symbol for the gene
of interest.}

\item{geom}{A string with the geom for plotting.}

\item{label}{A string with the name of the column from
\code{\link[SummarizedExperiment]{colData}} to color the points}

\item{facet}{A string with the name of the column from
\code{\link[SummarizedExperiment]{colData}} to separate the plot into facets.}

\item{dodge.width}{A numeric that adds dodge between the label categories.}

\item{assay}{String with the name of the assay for plotting.}
}
\value{
A ggplot object with a plot of the gene-wise copy number states.
}
\description{
Visualization for gene-wise copy number states
}
\details{
plotGeneCopy finds overlaps of the varbin scaffolds genomic ranges
which can be accessed with \code{\link[SummarizedExperiment]{rowRanges}}
with the genes from the assemblies of either hg19 or hg38. The genomic ranges
from hg19 comes from package \code{TxDb.Hsapiens.UCSC.hg19.knownGene} whereas
for hg38 package \code{TxDb.Hsapiens.UCSC.hg38.knownGene}.

If the argument geom is set to 'barplot' plotGeneCopy calculates the gene-wise
 frequencies of each copy number state for the selected genes across all of
 the cells. For this reason, geom 'barplot' can only be used with the argument
 assay set to 'integer'.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
plotGeneCopy(copykit_obj, genes = c("FHIT", "PTEN", "FOXO1", "BRCA1"))
plotGeneCopy(copykit_obj,
    genes = c("FHIT", "PTEN", "FOXO1", "BRCA1"),
    label = "subclones"
)
plotGeneCopy(copykit_obj,
    genes = c("FHIT", "PTEN", "FOXO1", "BRCA1"),
    label = "subclones", dodge.width = 0.8
)
}
\author{
Darlan Conterno Minussi
}
