% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(scCNA, order_cells = "graph_search")
}
\arguments{
\item{scCNA}{scCNA object.}
}
\value{
A heatmap visualization.
}
\description{
Plots a heatmap of the copy number data. Each row is a cell and colums represent genomic positions.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
