% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(scCNA, order_cells = "phylogeny", n_threads = 1)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{n_threads}{Number of threads used to calculate the distance. Passed to `amap::Dist`}
}
\value{
A heatmap visualization.
}
\description{
Plots a heatmap of the copy number data. Each row is a cell and colums represent genomic positions.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
