% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  scCNA,
  assay = "segment_ratios",
  order_cells = c("consensus_tree", "hclust", "phylogeny"),
  label = NULL,
  label_colors = NULL,
  group = NULL,
  consensus = FALSE,
  rounding_error = FALSE,
  genes = NULL,
  col = NULL,
  row_split = NULL,
  use_raster = TRUE,
  raster_quality = 2,
  n_threads = 1
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{assay}{String with the assay to pull data from to plot heatmap.}

\item{order_cells}{A string with the desired method to order the cells within}

\item{label}{A vector with the string names of the columns from
\code{\link[SummarizedExperiment]{colData}} for heatmap annotation.}

\item{label_colors}{A named list with colors for the label annotation.
Must match label length and have the same names as label}

\item{group}{with the names of the columns from
\code{\link[SummarizedExperiment]{colData}} to add a barplot with frequency
of the groups to a consensus heatmap.}

\item{consensus}{A boolean indicating if the consensus heatmap should be
plotted.}

\item{rounding_error}{A boolean indicating if the rounding error matrix
should be plotted.}

\item{genes}{A character vector with the HUGO symbol for genes to annotate
on the heatmap.}

\item{col}{A colorRamp2 vector that controls the color scale of the heatmap.
See \code{\link[ComplexHeatmap]{Heatmap}} or ComplexHeatmap online docs for
help.}

\item{row_split}{A string with the names of the columns from
\code{\link[SummarizedExperiment]{colData}} to split the heatmap.}

\item{use_raster}{Whether render the heatmap body as a raster image.
It helps to reduce file size when the matrix is huge. If number of rows or
columns is more than 2000, it is by default turned on. Note if cell_fun is
set, use_raster is enforced to be FALSE.
see \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{raster_quality}{A value larger than 1. Larger values increase the
file size.}

\item{n_threads}{Number of threads passed on to \code{runDistMat}.}
}
\value{
A \code{ComplexHeatmap} object with a heatmap of copy number data
where the columns are the genomic positions and each row is a cell.
}
\description{
Plots a heatmap of the copy number data.
Each row is a cell and colums represent genomic positions.
}
\details{
\itemize{
   \item{order_cells}: If order_cells argument is set to 'consensus_tree'
   \code{\link{plotHeatmap}} checks for the existence of a consensus matrix.
   From the consensus matrix, a minimum evolution tree is built and cells are
    ordered following the order of their respective groups from the tree.
    If order_cells is 'hclust' cells are ordered according to hierarchical
     clustering. 'hclust' calculation can be sped up by changing the parameter
     'n_threads' if you have more threads available to use. If order_cells
     is set to 'phylogeny' \code{\link{plotHeatmap}} will use the tree stored
     in the \code{\link{phylo}} slot of the scCNA object to order the cells.
     If the \code{\link{phylo}} is empty \code{\link{plotHeatmap}} will run
     \code{\link{runPhylo}} to generate the tree.

   \item{label}: A vector with the string names of the columns from
   \code{\link[SummarizedExperiment]{colData}} for heatmap annotation. The 'label'
   argument can take as many columns as desired as argument as long as they
   are elements from \code{\link[SummarizedExperiment]{colData}}.

   \item{label_colors}: A named list, list element names must match column
   names for \code{\link[SummarizedExperiment]{colData}} and list elements
   must match the number of items present in the columns provided in argument
    'label'. For example: to set colors for column 'outlier' containing
    elements 'TRUE' or 'FALSE' a valid input would be:
    'list(outlier = c('FALSE' = 'green', 'TRUE' = 'red))'.
     Default colors are provided for 'superclones', 'subclones', 'is_aneuploid',
     and 'outlier' that can be override with 'label_colors'.

   \item{rounding_error}: Must be used with assay = 'integer'.
   \code{plotHeatmap} will access the ploidies stored into colData(scCNA)$ploidy
    that are generated from \code{\link{calcInteger}} and scale rounded integer
     values to the segment means. Later this scaled matrix will be subtracted
     from the 'integer' assay from \code{\link{calcInteger}} and the resulting
      matrix from this subtraction will be plotted. Useful to visualize regions
      of high rounding error. Such regions can indicate issues with the ploidy
      scaling in use.

   \item{consensus}: If set to TRUE, \code{\link{plotHeatmap}} will search for
    the consensus matrix in the slot \code{\link{consensus}} and plot the
    resulting matrix. Labels annotations can be added with the argument 'label'.

}
}
\examples{
copykit_obj <- copykit_example_filtered()
set.seed(1000)
copykit_obj <- copykit_obj[, sample(200)]
copykit_obj <- findClusters(copykit_obj)
copykit_obj <- calcConsensus(copykit_obj)
copykit_obj <- runConsensusPhylo(copykit_obj)
colData(copykit_obj)$section <- stringr::str_extract(
    colData(copykit_obj)$sample,
    "(L[0-9]+L[0-9]+|L[0-9]+)"
)
plotHeatmap(copykit_obj, label = c("section", "subclones"))
}
\references{
Zuguang Gu, Roland Eils, Matthias Schlesner, Complex heatmaps
reveal patterns and correlations in multidimensional genomic data,
Bioinformatics, Volume 32, Issue 18, 15 September 2016, Pages 2847–2849,
https://doi.org/10.1093/bioinformatics/btw313
}
\seealso{
\code{\link{calcInteger}} For methods of obtaning the 'integer' assay.
}
\author{
Darlan Conterno Minussi
}
