% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(
  scCNA,
  order_cells = "hclust",
  label = NULL,
  label_colors = NULL,
  use_default_colors = TRUE,
  consensus = FALSE,
  row_split = NULL
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{order_cells}{Methods to order the cells within the heatmap. Accepted values are "phylogeny", "hclust", "graph_search". Defaults to "hclust".}

\item{label}{Character. Annotate heatmap by an element of metadata. Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}}

\item{label_colors}{List. Named list with colors for the label annotation. Must match label length}

\item{use_default_colors}{Boolean. Use the default colors when annotating superclones, subclones or filtering.}

\item{consensus}{Boolean. Indicates if the consensus heatmap should be plotted.}

\item{row_split}{Character. Element of the metadata to split the heatmap. Must have length = 1.}
}
\value{
A heatmap visualization.
}
\description{
Plots a heatmap of the copy number data. Each row is a cell and colums represent genomic positions.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
