% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{Plot metrics}
\usage{
plotMetrics(scCNA)
}
\arguments{
\item{scCNA}{scCNA object.}
}
\value{
A plot with the metrics available in the metadata

Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}
}
\description{
Provides a plot of the metrics to quickly assess the quality and overall look of the copy number data.
}
\examples{


}
\author{
Darlan Conterno Minussi
}
