% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{plotMetrics}
\alias{plotMetrics}
\title{plotMetrics}
\usage{
plotMetrics(scCNA, metric, label = NULL)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{metric}{A character indicating which elements of \code{colData()}
should be plotted.}

\item{label}{A character indicating which element of the \code{colData()} to
color the plots.}
}
\value{
A ggplot object with swarm plots of the selected metrics.
}
\description{
Plots swarm plots from selected elements of \code{\link[SummarizedExperiment{colData}]}.
}
\examples{


}
\author{
Darlan Conterno Minussi
}
