% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylo.R
\name{plotPhylo}
\alias{plotPhylo}
\title{Plot phylogenetic tree}
\usage{
plotPhylo(scCNA, label = NULL, label_colors = NULL, consensus = FALSE)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{label}{A string with the element of \code{\link[SummarizedExperiment]{colData}}.
to annotate the tips of the tree.}

\item{label_colors}{A named list with colors for the label annotation.
Must match label length}

\item{consensus}{A boolean indicating if the consensus phylogenetic tree
should be plotted.}
}
\value{
A phylogenetic tree visualization.
}
\description{
Plots a phylogenetic tree.
}
\examples{

}
\author{
Junke Wang
}
