% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylo.R
\name{plotPhylo}
\alias{plotPhylo}
\title{Plot phylogenetic tree}
\usage{
plotPhylo(scCNA, label = NULL, label_colors = NULL, consensus = FALSE)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{label}{Character. Annotate tree tip by one of the element of metadata.
Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}}

\item{label_colors}{List. Named list with colors for the label annotation.
Must match label length}

\item{consensus}{Boolean. Indicates if the consensus phylogenetic tree should be plotted. Default to FALSE.}
}
\value{
A phylogenetic tree visualization.
}
\description{
Plots a phylogenetic tree.
}
\examples{

}
\author{
Junke Wang
}
