% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatio.R
\name{plotRatio}
\alias{plotRatio}
\title{Plot ratio}
\usage{
plotRatio(scCNA, sample_name = NULL, interactively = FALSE)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{sample_name}{character vector with the name of the sample to be visualized}
}
\value{
Ratio plot from the selected sample.
}
\description{
plotRatioPlot allows for a visualization of the segment ratios together with the ratios.
It is helpful to observe the fit of the segmentation to the data.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
