% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSuggestedK.R
\name{plotSuggestedK}
\alias{plotSuggestedK}
\title{plotSuggestedK}
\usage{
plotSuggestedK(scCNA, geom = c("boxplot", "tile", "dotplot", "scatterplot"))
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{geom}{A character with the geom to be used for plotting.}
}
\value{
A ggplot2 object with the plot of different tested k values and their
jaccard similarity for each subclone
}
\description{
Uses the information from \code{\link{findSuggestedK}} to plot the values
of jaccard similarity from the tested k range on \code{\link{findSuggestedK}}.
}
\details{
\code{\link{plotSuggestedK}} access the \code{\link[S4Vectors]{metadata}}
element suggestedK_df that is saved to the scDNA object after running
\code{\link{findSuggestedK}}. The dataframe is used for plotting either a
heatmap, when the argument geom = 'tile', or a dotplot when argument geom =
'dotplot' or a boxplot when geom = 'boxplot'.

\itemize{
\item{geom = 'boxplot':} Plots a boxplot of the jaccard similarities across
all clusters detected in the grid search. The large red points represent
the mean jaccard similarity.

#' \item{geom = 'tile':} Plots a heatmap of the jaccard similarities across
all clusters detected in the grid search. The filling collors represent the
jaccard similarity value. Rows represent clusters and columns the k value.

#' \item{geom = 'dotplot':} Plots a dotplot of the jaccard similarities across
all clusters detected in the grid search. Where the size of the dots represent
the jaccard similarity for each assesed k value.

#' \item{geom = 'scatterplot':} Plots a scatterplot of the jaccard similarity
explained by the number of cells. Points are colored by subclone and lines
represent a linear regression across the points.

}
}
\examples{
copykit_obj <- copykit_example()
copykit_obj <- findNormalCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$is_normal == "FALSE"]
copykit_obj <- filterCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$filtered == "kept"]
copykit_obj <- runUmap(copykit_obj)
copykit_obj <- findSuggestedK(copykit_obj)
plotSuggestedK(copykit_obj)

}
