% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUmap.R
\name{plotUmap}
\alias{plotUmap}
\title{Plot UMAP embedding}
\usage{
plotUmap(scCNA)
}
\arguments{
\item{scCNA}{scCNA object.}
}
\value{
A reduced dimension representation with UMAP in the slot \code{reducedDim} from scCNA object. Access reduced dimensions slot with: \code{SummarizedExperiment::reducedDim(scCNA, 'umap')}. \code{plotUmap} searches for cluster information in the \code{SummarizedExperiment::colData()} metadata and colors the clusters according to that information.
}
\description{
Plots UMAP embedding stored in \code{reducedDims()} slot.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
