% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariableGenes.R
\name{plotVariableGenes}
\alias{plotVariableGenes}
\title{plotVariableGenes}
\usage{
plotVariableGenes(scCNA, n = 30)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{n}{A numeric defining how many variable genes will be plotted.}
}
\value{
A ggplot object with a plot of the variable genes detected.
}
\description{
Visualization for the most variable genes found with \code{findVariableGenes}.
}
\details{
\code{plotVariableGenes} plots the list of genes that was found
using \code{findVariableGenes}.
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findVariableGenes(copykit_obj,
    genes = c("FHIT", "PTEN", "FOXO1", "BRCA1")
)
plotVariableGenes(copykit_obj)
}
\seealso{
\code{\link{findVariableGenes}}
}
