% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariableGenes.R
\name{plotVariableGenes}
\alias{plotVariableGenes}
\title{plotVariableGenes}
\usage{
plotVariableGenes(scCNA, n = 30)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{n}{A numeric defining how many variable genes will be plotted.}
}
\value{

}
\description{
Visualization for the most variable genes found with \code{findVariableGenes}.
}
\details{
\code{plotVariableGenes} plots the list of genes that was found
using \code{findVariableGenes}.
}
\examples{
copykit_obj <- copykit_example()
copykit_obj <- findNormalCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$is_normal == "FALSE"]
copykit_obj <- filterCells(copykit_obj)
copykit_obj <- copykit_obj[,colData(copykit_obj)$filtered == "kept"]
copykit_obj <- findVariableGenes(copykit_obj)
plotVariableGenes(copykit_obj)

}
\seealso{
\code{\link{findVariableGenes}}
}
