% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCNAVarbin.R
\name{readCNAVarbin}
\alias{readCNAVarbin}
\title{Load data from Copy Number Experiment}
\usage{
readCNAVarbin(dir, remove_Y = FALSE)
}
\arguments{
\item{dir}{A path for the output of the copy number pipeline.}
}
\value{
A S4 class object containing the segment ratios for each bin (rows) and each sample (columns).
}
\description{
Creates an S4 class scCNA from the output directory of the copy number pipeline.
}
\examples{

}
