% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runConsensusPhylo.R
\name{runConsensusPhylo}
\alias{runConsensusPhylo}
\title{runConsensusPhylo}
\usage{
runConsensusPhylo(scCNA, root = c("mrca", "neutral"))
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{root}{A string indicating how to root the consensus tree.}
}
\value{

}
\description{
Runs a minimal evolution tree algorithm for the consensus data frame
}
