% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runConsensusPhylo.R
\name{runConsensusPhylo}
\alias{runConsensusPhylo}
\title{runConsensusPhylo}
\usage{
runConsensusPhylo(scCNA, root = c("mrca", "neutral", "user"), root_user = NULL)
}
\arguments{
\item{scCNA}{The scCNA object.}

\item{root}{A string indicating how to root the consensus tree.}

\item{root_user}{A numeric with the vector to be used as root of the tree if
\code{root} is set to 'user'. Must have the same length as the number of bins
of the genome scaffold.}
}
\value{
A phylo object with a consensus tree stored in the consensusPhylo slot
of the CopyKit object.
}
\description{
Runs a minimal evolution tree algorithm for the consensus data frame
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- findClusters(copykit_obj)
copykit_obj <- calcConsensus(copykit_obj)
copykit_obj <- runConsensusPhylo(copykit_obj)
plotPhylo(copykit_obj, consensus = TRUE, label = "subclones")
}
