% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetrics.R
\name{runMetrics}
\alias{runMetrics}
\title{Run metrics}
\usage{
runMetrics(scCNA, BPPARAM = bpparam())
}
\arguments{
\item{scCNA}{scCNA object.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
Adds a the metrics to the scCNA metadata. Those metrics can be used for subsetting the data if desired

Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}
}
\description{
Calculates the RMSE and the breakpoint counts for each cell.
Stores it in the metadata that can be accessed with \code{SummarizedExperiment::colData(scCNA)}
Results can be visualized with \code{plotMetrics()}
}
\examples{


}
\author{
Darlan Conterno Minussi
}
