% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetrics.R
\name{runMetrics}
\alias{runMetrics}
\title{Run metrics}
\usage{
runMetrics(scCNA, n_threads = parallel::detectCores()/4)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{n_threads}{Number of threads used to calculate the distance matrix. Passed to `parallel::mclapply`. As default it uses 1/4 of the detected cores available.}
}
\value{
Adds a the metrics to the scCNA metadata. Those metrics can be used for subsetting the data if desired

Metadata can be accessed with \code{SummarizedExperiment::colData(scCNA)}
}
\description{
Provides some metrics to quickly assess the quality and overall look of the copy number data.
}
\examples{

}
\author{
Darlan Conterno Minussi
}
