% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetrics.R
\name{runMetrics}
\alias{runMetrics}
\title{Run metrics}
\usage{
runMetrics(scCNA, BPPARAM = bpparam())
}
\arguments{
\item{scCNA}{scCNA object.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
Adds columns 'overdispersion' and 'breakpoint_count' to
 \code{\link[SummarizedExperiment]{colData}}.
}
\description{
Calculates the overdispersion and the breakpoint counts for each cell.
}
\details{
Adds the metrics to the scCNA \code{\link[SummarizedExperiment]{colData}}.
 Those metrics can be used for subsetting the data if desired.
 results can be visualized with \code{\link{plotMetrics}}.
}
\examples{
copykit_obj <- mock_bincounts()
copykit_obj <- runVst(copykit_obj)
copykit_obj <- calcRatios(copykit_obj)
copykit_obj <- runSegmentation(copykit_obj)
copykit_obj <- runMetrics(copykit_obj)
}
\author{
Darlan Conterno Minussi
}
