% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run Segmentation}
\usage{
runSegmentation(
  scCNA,
  method = "CBS",
  seed = 17,
  name = "segment_ratios",
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{The scCNA object}

\item{method}{Character. Segmentation method of choice.}

\item{seed}{Numeric. Set seed for CBS segmentation permutation reproducibility.}

\item{name}{Character. Target slot for the resulting segment ratios.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
The segment profile for all cells inside the scCNA object. Can be retrieved with \code{copykit::segment_ratios()}
}
\description{
Runs a segmentation algorithm using the ratio data.
}
