% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run Segmentation}
\usage{
runSegmentation(
  scCNA,
  method = c("CBS", "multipcf"),
  seed = 17,
  alpha = 1e-10,
  gamma = 40,
  undo.splits = "prune",
  name = "segment_ratios",
  BPPARAM = bpparam()
)
}
\arguments{
\item{scCNA}{The scCNA object}

\item{method}{A character with the segmentation method of choice.}

\item{seed}{Numeric. Set seed for CBS segmentation permutation reproducibility}

\item{alpha}{A numeric with the. significance levels for the test to accept
change-points for CBS segmentation. See \code{\link[DNAcopy]{segment}}.}

\item{gamma}{A numeric passed on to 'multipcf' segmentation. Penalty for each
discontinuity in the curve, default is 40. See \code{\link[copynumber]{multipcf}}.}

\item{undo.splits}{A character string specifying how change-points are to be
undone, if at all. Default is "none". Other choices are "prune", which uses
 a sum of squares criterion, and "sdundo", which undoes splits that are not
 at least this many SDs apart. See \code{\link[DNAcopy]{segment}}}

\item{name}{Character. Target slot for the resulting segment ratios.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} specifying how the function
should be parallelized.}
}
\value{
The segment profile for all cells inside the scCNA object.
}
\description{
Runs a segmentation algorithm using the ratio data.
}
\details{
\itemize{

   \item{CBS:} #' \code{runSegmentation} Fits a piece-wise constant function
   to the transformed the smoothed bin counts. Bin counts are smoothed with
   \code{\link[DNAcopy]{smooth.CNA}} using the Circular Binary Segmentation
   (CBS) algorithm from \code{\link[DNAcopy]{segment}} with default it applies
   undo.prune with value of 0.05.

   \item{multipcf:} Performs the joint segmentation from the \code{copynumber}
   package using the \code{\link[copynumber]{multipcf}} function. By fitting
   piecewise constant curves with common breakpoints for all samples.


}


The resulting segment means are further refined with MergeLevels to join
adjacent segments with non-significant differences in segmented means.
}
\examples{
\dontrun{
copykit_obj <- runSegmentation(copykit_obj)
}

}
