% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSegmentation.R
\name{runSegmentation}
\alias{runSegmentation}
\title{Run Segmentation}
\usage{
runSegmentation(
  scCNA,
  method = "CBS",
  genome = "hg38",
  seed = 17,
  n_threads = parallel::detectCores()/4
)
}
\arguments{
\item{scCNA}{The scCNA object}

\item{method}{Character. Segmentation method of choice.}

\item{genome}{Character. Genome assembly to be used, current accepted "hg19" or "hg38".}

\item{seed}{Numeric. Set seed for CBS segmentation permutaton reproducibility.}

\item{n_threads}{Number of threads used to calculate the distance matrix. Passed to `parallel::mclapply`. As default it uses 1/4 of the detected cores available.}
}
\value{
The segment profile for all cells inside the scCNA object. Can be retrieved with \code{copykit::segment_ratios()}
}
\description{
Runs a segmentation algorithm using the ratio data.
}
