% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmap.R
\name{runUmap}
\alias{runUmap}
\title{Creates UMAP embedding}
\usage{
runUmap(
  scCNA,
  assay = "logr",
  seed = 17,
  min_dist = 0,
  n_neighbors = 50,
  name = "umap",
  ncomponents = 2,
  ...
)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{assay}{String with the name of the assay to pull data from to make the
embedding.}

\item{seed}{Sets a seed for the pseudorandom number generator.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where nearby
points on the manifold are drawn closer together, while larger values will
result on a more even dispersal of points. The value should be set relative
to the spread value, which determines the scale at which embedded points
will be spread out. See \code{\link[uwot]{umap}}.}

\item{n_neighbors}{The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation.
Larger values result in more global views of the manifold,
while smaller values result in more local data being preserved.
In general values should be in the range 2 to 100.
See \code{\link[uwot]{umap}}.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{ncomponents}{The dimension of the space to embed into. See
\code{\link[uwot]{umap}}.}

\item{...}{Additional parameters passed to \code{\link[uwot]{umap}}.}
}
\value{
A reduced dimension representation with UMAP in the slot
 \code{reducedDim} from scCNA object. Access reduced dimensions slot with:
 \code{reducedDim(scCNA, 'umap', withDimnames = FALSE)}
}
\description{
Creates a umap embedding using the package uwot from the segment ratios
 values
}
\examples{
copykit_obj <- copykit_example_filtered()
copykit_obj <- runUmap(copykit_obj)
}
\author{
Darlan Conterno Minussi
}
