% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUmap.R
\name{runUmap}
\alias{runUmap}
\title{Creates UMAP embedding}
\usage{
runUmap(scCNA, seed = 17, ...)
}
\arguments{
\item{scCNA}{scCNA object.}

\item{seed}{Sets a seed for the pseudorandom number generator.}

\item{...}{Additional parameters passed to \code{uwot::umap}.}
}
\value{
A reduced dimension representation with UMAP in the slot \code{reducedDim} from scCNA object. Access reduced dimensions slot with: \code{SingleCellExperiment::reducedDims(scCNA, 'umap')}
}
\description{
Creates a umap embedding using the package uwot from the segment ratios values
}
\author{
Darlan Conterno Minussi
}
