% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVst.R
\name{runVst}
\alias{runVst}
\title{Variance Stabilizing Transformation}
\usage{
runVst(scCNA, transformation = "ft")
}
\arguments{
\item{scCNA}{The scCNA object}

\item{transformation}{Character. Transformation to be performed, available options are 'log' or 'ft'}
}
\value{
A slot into the scCNA object containing the variance stabilized matrix.
}
\description{
Performs variance stabilization by using the Freeman-Tukey transformation
}
