% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVst.R
\name{runVst}
\alias{runVst}
\title{Variance Stabilizing Transformation}
\usage{
runVst(scCNA, transformation = "ft")
}
\arguments{
\item{scCNA}{The scCNA object}

\item{transformation}{Character. Transformation to be performed, available options are 'log' or 'ft'}
}
\value{
A slot into the scCNA object containing the variance stabilized matrix.
}
\description{
Performs variance stabilization by using the Freeman-Tukey transformation
}
\details{
\code{runVst} performs variance stabilization to reduce the overdispersion
from the negative binomial distribution nature of the bin counts and reduce
technical bias. The argument \code{vst} controls the choice of the transformation
allowing either the Freeman-Tukey transformation by using the option 'ft' (recommended)
or a logarithmic transformation with the option 'log'. Using a 'log' transformation
may result in long segmentation times for a few cells with large breakpoint counts.
}
