% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVst.R
\name{runVst}
\alias{runVst}
\title{Variance Stabilizing Transformation}
\usage{
runVst(scCNA, transformation = c("ft", "log"))
}
\arguments{
\item{scCNA}{The scCNA object}

\item{transformation}{A character indicating the variance stabilization
transformation to be performed. See \link{runVst} details.}
}
\value{
A slot into the scCNA object containing the variance stabilized matrix.
}
\description{
Performs variance stabilization transformation of the bin counts
}
\details{
\code{runVst} performs variance stabilization to reduce the overdispersion
from the negative binomial distribution nature of the bin counts and reduce
technical bias. The argument \code{vst} controls the choice of the transformation
allowing either the Freeman-Tukey transformation by using the option 'ft' (recommended)
or a logarithmic transformation with the option 'log'. Using a 'log' transformation
may result in long segmentation times for a few cells with large breakpoint counts.
}
\examples{
\dontrun{
copykit_obj <- runVst(copykit_obj)
}

}
\references{
Freeman, M. F.; Tukey, J. W. (1950), "Transformations related to the angular
and the square root", The Annals of Mathematical Statistics,
21 (4), pp. 607–611, doi:10.1214/aoms/1177729756, JSTOR 2236611
}
