% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{scCNA-class}
\alias{scCNA-class}
\alias{.scCNA}
\title{@export
 @import methods
 @importClassesFrom SummarizedExperiment RangedSummarizedExperiment SingleCellExperiment
 @importClassesFrom S4Vectors DataFrame SimpleList}
\description{
@export
 @import methods
 @importClassesFrom SummarizedExperiment RangedSummarizedExperiment SingleCellExperiment
 @importClassesFrom S4Vectors DataFrame SimpleList
}
